move image generation from dc module to logic
